pyproject.toml
[project]
name = "qt-py-logs"
version = "0.1.1"
description = "python client for https://github.com/ausward/QTLogs"
authors = [
    {name = "Austin",email = "71802516+ausward@users.noreply.github.com"}
]
license = {text = "CC NC SA"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
]

[tool.poetry.urls]
"Homepage" = "https://github.com/ausward/qt-py-logs"
"Repository" = "https://github.com/ausward/qt-py-logs"
"Bug Tracker" = "https://github.com/ausward/qt-py-logs/issues"


[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0,<8.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
